Updated to the new PokeEmerald-Expansion Pokedex Plus, Re-enabled the displaying of tera types in  the summary menu, and added the ability to change the tera type
Updated to the new PokeEmerald-Expansion Pokedex Plus, which provides useful information
such as base stats, moves, formes and more. Re-enabled the displaying of tera types in
the summary menu, and added the ability to change the tera type of a Pokémon by using
a single tera shard on them.